Resolved template instantiation mismatch
git-svn-id: http://usg.lofar.org/svn/code/trunk@1924 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/CR-Tools/implement/Analysis/test/tppfimplement.cc
@@ -192,7 +192,7 @@ int test_ppfimplements ()
     }
     
     Vector<Double> freq_vector_fft( 513,0.0 );
-    Vector<uint> freq_vector_shortened( 510,0.0 );
+    Vector<Double> freq_vector_shortened( 510,0.0 );
     
     for( uint g=0; g<512; g++ ){
